Top level http (#83)
* Modularized http and made http a top level module

Modularized the http directory by separating related functions into
different scripts. Also removed EventClass and kept a singular
CloudEvent. Finally, CloudEvent.__repr__ was refactored such that it
doesn't depend on external methods.


* renamed requests.py to http_methods


* lint fixes
