+++ cloudevents/http/__init__.py
@@ -11,27 +11,13 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
-from flask import Flask, request
+import json
+import typing
 
-from cloudevents.http import from_http
-
-app = Flask(__name__)
-
-
-# create an endpoint at http://localhost:/3000/
-@app.route("/", methods=["POST"])
-def home():
-    # create a CloudEvent
-    event = from_http(request.get_data(), request.headers)
-
-    # you can access cloudevent fields as seen below
-    print(
-        f"Found {event['id']} from {event['source']} with type "
-        f"{event['type']} and specversion {event['specversion']}"
-    )
-
-    return "", 204
-
-
-if __name__ == "__main__":
-    app.run(port=3000)
+from cloudevents.http.event import CloudEvent
+from cloudevents.http.http_methods import (
+    from_http,
+    to_binary_http,
+    to_structured_http,
+)
+from cloudevents.http.json_methods import from_json, to_json
