+++ cloudevents/http/event.py
@@ -13,39 +13,13 @@
 #    under the License.
 
 import datetime
-import json
 import typing
 import uuid
 
-from cloudevents.sdk import converters, marshaller, types
-from cloudevents.sdk.converters import is_binary
-from cloudevents.sdk.event import v1, v03
-from cloudevents.sdk.marshaller import HTTPMarshaller
+from cloudevents.http.mappings import _required_by_version
 
 
-def default_marshaller(content: any):
-    if content is None or len(content) == 0:
-        return None
-    try:
-        return json.dumps(content)
-    except TypeError:
-        return content
-
-
-_marshaller_by_format = {
-    converters.TypeStructured: lambda x: x,
-    converters.TypeBinary: default_marshaller,
-}
-
-_obj_by_version = {"1.0": v1.Event, "0.3": v03.Event}
-
-_required_by_version = {
-    "1.0": v1.Event._ce_required_fields,
-    "0.3": v03.Event._ce_required_fields,
-}
-
-
-class EventClass:
+class CloudEvent:
     """
     Python-friendly cloudevent class supporting v1 events
     Supports both binary and structured mode CloudEvents
@@ -114,71 +88,5 @@ class EventClass:
     def __contains__(self, key):
         return key in self._attributes
 
-
-def _to_http(
-    event: EventClass,
-    format: str = converters.TypeStructured,
-    data_marshaller: types.MarshallerType = None,
-) -> (dict, typing.Union[bytes, str]):
-    """
-    Returns a tuple of HTTP headers/body dicts representing this cloudevent
-
-    :param format: constant specifying an encoding format
-    :type format: str
-    :param data_marshaller: Callable function to cast event.data into
-        either a string or bytes
-    :type data_marshaller: types.MarshallerType
-    :returns: (http_headers: dict, http_body: bytes or str)
-    """
-    if data_marshaller is None:
-        data_marshaller = _marshaller_by_format[format]
-
-    if event._attributes["specversion"] not in _obj_by_version:
-        raise ValueError(
-            f"Unsupported specversion: {event._attributes['specversion']}"
-        )
-
-    event_handler = _obj_by_version[event._attributes["specversion"]]()
-    for k, v in event._attributes.items():
-        event_handler.Set(k, v)
-    event_handler.data = event.data
-
-    return marshaller.NewDefaultHTTPMarshaller().ToRequest(
-        event_handler, format, data_marshaller=data_marshaller
-    )
-
-
-def to_structured_http(
-    event: EventClass, data_marshaller: types.MarshallerType = None,
-) -> (dict, typing.Union[bytes, str]):
-    """
-    Returns a tuple of HTTP headers/body dicts representing this cloudevent
-
-    :param event: CloudEvent to cast into http data
-    :type event: CloudEvent
-    :param data_marshaller: Callable function to cast event.data into
-        either a string or bytes
-    :type data_marshaller: types.MarshallerType
-    :returns: (http_headers: dict, http_body: bytes or str)
-    """
-    return _to_http(event=event, data_marshaller=data_marshaller)
-
-
-def to_binary_http(
-    event: EventClass, data_marshaller: types.MarshallerType = None,
-) -> (dict, typing.Union[bytes, str]):
-    """
-    Returns a tuple of HTTP headers/body dicts representing this cloudevent
-
-    :param event: CloudEvent to cast into http data
-    :type event: CloudEvent
-    :param data_marshaller: Callable function to cast event.data into
-        either a string or bytes
-    :type data_marshaller: types.UnmarshallerType
-    :returns: (http_headers: dict, http_body: bytes or str)
-    """
-    return _to_http(
-        event=event,
-        format=converters.TypeBinary,
-        data_marshaller=data_marshaller,
-    )
+    def __repr__(self):
+        return str({"attributes": self._attributes, "data": self.data})
