+++ samples/http-json-cloudevents/client.py
@@ -16,7 +16,7 @@ import sys
 
 import requests
 
-from cloudevents.sdk.http import CloudEvent, to_binary_http, to_structured_http
+from cloudevents.http import CloudEvent, to_binary_http, to_structured_http
 
 
 def send_binary_cloud_event(url):
